LedgerMonolith refactor (1 cache, 1 controller) (#411)
* Merge most recent version changes into ledgermonolith source code

* Update ledgermonolith tests to actually pass on build

* Minor fixes in ledgermonolith-service (#408)

* Fix pubkey file creation to avoid crashes after 2nd reboot

* Allow getting GCS_BUCKET from env var

* Update ledgermonolith README with commands to run the full BoA app (#402)

* Update ledgermonolith README with commands to run the full BoA app

* Add clarifying error messages for GCS_BUCKET when building the monolith



* Monolith refactor

* Merge caches

* Combine txnhistory and balancereader caches

* Cache workaround

* Cleanup

* Fix ProcessTransaction bug

* Adds monolith to checkstyle

* Cleanup

* Fix pubkey file creation to avoid crashes after 2nd reboot

* Allow getting GCS_BUCKET from env var

* ledgermonolith: introduce script for one-time install

Currently, the creation of ledgermonolith-service VM is
tailored for faster demo, by using startup-script.sh which
covers both installation activities and service startup.
This approach, howerver, has two main drawbacks as the script:
- installs / updates binaries on every VM start
- initializes / generates DB on every VM start

New version of the script:
1. Checks if service has been already installed
2. If not:
- pre-requisites
- downloads binaries
- initializes DB
- installs enables systemd service to run ledgermonolith app

* Remove sensitive log statement from LedgerReaderCache
