+++ src/ledgermonolith/checkstyle.xml
@@ -124,9 +124,15 @@
 
         <!-- Checks for Size Violations.                    -->
         <!-- See http://checkstyle.sf.net/config_sizes.html -->
-        <module name="LineLength"/>
+        <module name="LineLength">
+            <property name="max" value="100"/>
+        </module>
         <module name="MethodLength"/>
-        <module name="ParameterNumber"/>
+         <module name="ParameterNumber">
+          <property name="max" value="10"/>
+          <property name="tokens" value="METHOD_DEF"/>
+        </module>
+
 
 
         <!-- Checks for whitespace                               -->
