+++ src/ledgermonolith/init/ledgermonolith.env
@@ -1,9 +1,8 @@
-VERSION="v0.1.0"
+VERSION="v0.3.1"
 PORT="8080"
 JVM_OPTS="-XX:+UnlockExperimentalVMOptions -XX:+UseCGroupMemoryLimitForHeap"
 LOCAL_ROUTING_NUM="883745000"
 PUB_KEY_PATH="/opt/monolith/publickey"
-BALANCES_API_ADDR="127.0.0.1:8080"
 POSTGRES_DB="postgresdb"
 POSTGRES_USER="postgres"
 POSTGRES_PASSWORD="password"
@@ -11,3 +10,5 @@ SPRING_DATASOURCE_URL="jdbc:postgresql://127.0.0.1:5432/postgresdb"
 SPRING_DATASOURCE_USERNAME="postgres"
 SPRING_DATASOURCE_PASSWORD="password"
 USE_DEMO_DATA="True"
+ENABLE_TRACING="false"
+ENABLE_METRICS="false"
