--- src/ledgermonolith/kubernetes-manifests/.gitignore
@@ -1,1 +1,0 @@
-config.yaml
