--- src/ledgermonolith/kubernetes-manifests/contacts.yaml
@@ -1,82 +1,0 @@
-# Copyright 2019 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: contacts
-spec:
-  selector:
-    matchLabels:
-      app: contacts
-  template:
-    metadata:
-      labels:
-        app: contacts
-    spec:
-      terminationGracePeriodSeconds: 5
-      containers:
-      - name: contacts
-        image: gcr.io/bank-of-anthos/contacts:latest
-        volumeMounts:
-        - name: publickey
-          mountPath: "/root/.ssh"
-          readOnly: true
-        env:
-        - name: VERSION
-          value: "v0.2.0"
-        - name: PORT
-          value: "8080"
-        # Valid levels are debug, info, warning, error, critical.
-        # If no valid level is set, will default to info.
-        - name: LOG_LEVEL
-          value: "info"
-        envFrom:
-        - configMapRef:
-            name: environment-config
-        - configMapRef:
-            name: accounts-db-config
-        resources:
-          requests:
-            cpu: 100m
-            memory: 64Mi
-          limits:
-            cpu: 500m
-            memory: 256Mi
-        readinessProbe:
-          httpGet:
-            path: /ready
-            port: 8080
-          initialDelaySeconds: 5
-          periodSeconds: 5
-      volumes:
-      - name: publickey
-        secret:
-          secretName: jwt-key
-          items:
-          - key: jwtRS256.key.pub
-            path: publickey
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: contacts
-spec:
-  type: ClusterIP
-  selector:
-    app: contacts
-  ports:
-  - name: http
-    port: 8080
-    targetPort: 8080
