--- src/ledgermonolith/kubernetes-manifests/userservice.yaml
@@ -1,92 +1,0 @@
-# Copyright 2019 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: userservice
-spec:
-  selector:
-    matchLabels:
-      app: userservice
-  template:
-    metadata:
-      labels:
-        app: userservice
-    spec:
-      terminationGracePeriodSeconds: 5
-      containers:
-      - name: userservice
-        image: gcr.io/bank-of-anthos/userservice:latest
-        volumeMounts:
-        - name: keys
-          mountPath: "/root/.ssh"
-          readOnly: true
-        ports:
-        - name: http-server
-          containerPort: 8080
-        env:
-        - name: VERSION
-          value: "v0.2.0"
-        - name: PORT
-          value: "8080"
-        - name: TOKEN_EXPIRY_SECONDS
-          value: "3600"
-        - name: PRIV_KEY_PATH
-          value: "/root/.ssh/privatekey"
-        # Valid levels are debug, info, warning, error, critical. If no valid level is set, gunicorn will default to info.
-        - name: LOG_LEVEL
-          value: "info"
-        envFrom:
-        - configMapRef:
-            name: environment-config
-        - configMapRef:
-            name: accounts-db-config
-        readinessProbe:
-          httpGet:
-            path: /ready
-            port: 8080
-          initialDelaySeconds: 5
-          periodSeconds: 5
-        resources:
-          requests:
-            cpu: 100m
-            memory: 64Mi
-          limits:
-            cpu: 500m
-            memory: 256Mi
-      volumes:
-      - name: keys
-        secret:
-          secretName: jwt-key
-          items:
-          - key: jwtRS256.key
-            path: privatekey
-          - key: jwtRS256.key.pub
-            path: publickey
-
-
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: userservice
-spec:
-  type: ClusterIP
-  selector:
-    app: userservice
-  ports:
-  - name: http
-    port: 8080
-    targetPort: 8080
