+++ src/ledgermonolith/src/main/resources/application.properties
@@ -9,6 +9,19 @@ logging.level.com.zaxxer.hikari.pool.HikariPool=off
 logging.level.com.zaxxer.hikari.pool.ProxyConnection=error
 hibernate.jdbc.lob.non_contextual_creation=true
 spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults=false
+spring.main.allow-bean-definition-overriding=true
+
 
 spring.jpa.database=postgresql
 spring.datasource.platform=postgres
+
+# tracing on/off toggle
+spring.sleuth.enabled = ${ENABLE_TRACING}
+# if tracing enabled, configuration options:
+spring.sleuth.sampler.probability=1.0
+spring.sleuth.web.skipPattern=(^cleanup.*|.+favicon.*)
+spring.sleuth.log.slf4j.enabled=false
+# enables all tomcat metrics to be recorded
+server.tomcat.mbeanregistry.enabled=true
+# enables jpa metrics to be recorded
+spring.jpa.properties.hibernate.generate_statistics=true
